[SPARK-31127][ML] Implement abstract Selector
